消除 Linux 编译警告

+++ fibjs/include/XmlDataImpl.h
@@ -43,7 +43,7 @@ public:
 
     result_t substringData(int32_t offset, int32_t count, std::string &retVal)
     {
-        if (offset < 0 || offset > m_data.length())
+        if (offset < 0 || offset > (int32_t)m_data.length())
             return CALL_E_BADINDEX;
         retVal = m_data.substr(offset, count);
         return 0;
@@ -57,7 +57,7 @@ public:
 
     result_t insertData(int32_t offset, const char *arg)
     {
-        if (offset < 0 || offset > m_data.length())
+        if (offset < 0 || offset > (int32_t)m_data.length())
             return CALL_E_BADINDEX;
 
         m_data = m_data.substr(0, offset) + arg + m_data.substr(offset);
@@ -66,10 +66,10 @@ public:
 
     result_t deleteData(int32_t offset, int32_t count)
     {
-        if (offset < 0 || offset > m_data.length())
+        if (offset < 0 || offset > (int32_t)m_data.length())
             return CALL_E_BADINDEX;
 
-        if (offset + count > m_data.length())
+        if (offset + count > (int32_t)m_data.length())
             count = (int32_t)m_data.length() - offset;
 
         m_data = m_data.substr(0, offset) + m_data.substr(offset + count);
@@ -78,10 +78,10 @@ public:
 
     result_t replaceData(int32_t offset, int32_t count, const char *arg)
     {
-        if (offset < 0 || offset > m_data.length())
+        if (offset < 0 || offset > (int32_t)m_data.length())
             return CALL_E_BADINDEX;
 
-        if (offset + count > m_data.length())
+        if (offset + count > (int32_t)m_data.length())
             count = (int32_t)m_data.length() - offset;
 
         m_data = m_data.substr(0, offset) + arg + m_data.substr(offset + count);
@@ -90,7 +90,7 @@ public:
 
     result_t splitText(int32_t offset, std::string &retVal)
     {
-        if (offset < 0 || offset > m_data.length())
+        if (offset < 0 || offset > (int32_t)m_data.length())
             return CALL_E_BADINDEX;
 
         retVal = m_data.substr(offset);
